chore : Home -> PreviewScreen으로 이동
