+++ src/navigations/BottomTabs.tsx
@@ -1,6 +1,7 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
 import Home from '@/screens/home/Home'
-import LoginScreen from '@/screens/register/LoginScreen'
+import PreviewScreen from '@/screens/previewScreen/PreviewScreen'
+import LoginScreen from '@/screens/register/login/LoginScreen'
 
 const Tab = createBottomTabNavigator()
 
@@ -8,6 +9,7 @@ const BottomTabs = () => {
   return (
     <Tab.Navigator>
       <Tab.Screen name="Home" component={Home} />
+      <Tab.Screen name="PreviewScreen" component={PreviewScreen} />
       <Tab.Screen name="LoginScreen" component={LoginScreen} />
     </Tab.Navigator>
   )
