--- src/screens/home/Home.style.ts
@@ -1,37 +1,0 @@
-import styled from 'styled-components/native'
-
-export const Container = styled.View`
-  flex: 1;
-  padding-left: 20px;
-  padding-right: 20px;
-  align-items: center;
-`
-export const ImageArea = styled.View`
-  width: 100%;
-  height: 292px;
-  background-color: gray;
-  border-radius: 12px;
-`
-export const TextArea = styled.View`
-  margin-top: 40px;
-  margin-bottom: 40px;
-  align-items: center;
-`
-export const BoldText = styled.Text`
-  font-size: 24px;
-  font-weight: 800;
-  color: #32343e;
-  margin-bottom: 20px;
-`
-export const NomalText = styled.Text`
-  font-size: 16px;
-  font-weight: 400;
-  color: #32343e;
-`
-export const ButtonWrapView = styled.View`
-  margin-top: 20px;
-  align-items: center;
-`
-export const SkipText = styled.Text`
-  margin-top: 30px;
-`
